datax-master/requirements.txt
Django==2.1.7
django-crispy-forms==1.7.2
django-recaptcha==2.0.2
django-registration==3.0
social-auth-app-django==3.1.0
shortuuid==0.5.0
psycopg2-binary==2.7.7